Add tap intercept view instead of tap GR
In iOS 6+, gesture recognizers were fiddled with to allow them to play
nice with buttons. This means that the tap gesture recognizer was
letting taps fall through to the content view controller, leading to
some mildly annoying behavior.

The fix is to add a touch intercepting view to the contentView that
already has the tap gesture recognizer added to it. This view will
intercept all touch events, keeping them from falling through to the
content view controller.