Remove explicit view appearance calls
These aren't actually needed, and lead to odd behavior when changing
content views (among other times). By removing these calls and letting
`IAMenuController` forward the appearance calls to the children, we can
keep these duplicate calls to a minimum.

The downside here is that the `contentViewController` now doesn't
receive any notifications when the menu opens or closes. If you need
this functionality, you can subscribe to the menu state notifications
that fire.