Change fix for double appearance forwarding
Unfortunately a recently discovered side-effect of overriding
shouldAutomaticallyForwardAppearanceMethods on IAMenuController is that
child view controllers stop receiving appearance methods when modal
controllers are dismissed on top of them. Other appearance methods seem
to be getting forwarded fine, but as a workaround for this, since the
podspec for this requires iOS6 and we can expect automatic appearance
forwarding to always be enabled, we should be able to just rely on it
when subviews are added/removed instead of the manual viewWillDisappear/
viewDidDisappear/viewDidAppear calls being made.